+++ .gitignore
@@ -16,4 +16,3 @@ integration-tests/browser/dist/bundle.js
 integration-tests/data/standalone-0/db/
 integration-tests/data/standalone-0/p2p/
 integration-tests/data/bootstrap-0/p2p/
-
