+++ rpc/buf
@@ -1,0 +1,1 @@
+{"error_string":"runtime error: invalid memory address or nil pointer dereference","level":"info","method_string":"AddOrders","msg":"RPC method handler crashed","myPeerID":"16Uiu2HAmPowPYpMpYz9XKLxmrJZd7cTGrXKhKYhhW7bDNssJTe7u","stackTrace_string":"goroutine 92 [running]:\nruntime/debug.Stack(0x49f30c0, 0xc0001caf68, 0x4b3ce13)\n\t/usr/local/Cellar/go/1.13.1/libexec/src/runtime/debug/stack.go:24 +0x9d\nmain.(*rpcHandler).AddOrders.func1(0xc0001cb8c8)\n\t/Users/alextowle/go/src/github.com/0xProject/0x-mesh/cmd/mesh/rpc_handler.go:112 +0x1ba\npanic(0x49f39c0, 0x560cf80)\n\t/usr/local/Cellar/go/1.13.1/libexec/src/runtime/panic.go:679 +0x1b2\ngithub.com/0xProject/0x-mesh/p2p.(*Node).Send(...)\n\t/Users/alextowle/go/src/github.com/0xProject/0x-mesh/p2p/node.go:551\ngithub.com/0xProject/0x-mesh/core.(*App).shareOrder(0xc00017e8c0, 0xc0001b81e0, 0xc0001cb4f8, 0x1)\n\t/Users/alextowle/go/src/github.com/0xProject/0x-mesh/core/core.go:758 +0xc4\ngithub.com/0xProject/0x-mesh/core.(*App).AddOrders(0xc00017e8c0, 0xc0004a81c0, 0x1, 0x4, 0x1, 0x49d5a40, 0xc0004825a0, 0xc000401758)\n\t/Users/alextowle/go/src/github.com/0xProject/0x-mesh/core/core.go:745 +0x1221\nmain.(*rpcHandler).AddOrders(0xc0000de048, 0xc0004a81c0, 0x1, 0x4, 0x1, 0x0, 0x0, 0x0)\n\t/Users/alextowle/go/src/github.com/0xProject/0x-mesh/cmd/mesh/rpc_handler.go:117 +0x26e\ngithub.com/0xProject/0x-mesh/rpc.(*rpcService).AddOrders(0xc000033000, 0xc0004a81c0, 0x1, 0x4, 0x0, 0x0, 0x0, 0x0)\n\t/Users/alextowle/go/src/github.com/0xProject/0x-mesh/rpc/service.go:132 +0x66\nreflect.Value.call(0xc0003ea000, 0xc0000e2070, 0x13, 0x4b3aa96, 0x4, 0xc0000c2ea0, 0x3, 0x4, 0xc000401bc8, 0x404667c, ...)\n\t/usr/local/Cellar/go/1.13.1/libexec/src/reflect/value.go:460 +0x5f6\nreflect.Value.Call(0xc0003ea000, 0xc0000e2070, 0x13, 0xc0000c2ea0, 0x3, 0x4, 0x2, 0x2, 0x2)\n\t/usr/local/Cellar/go/1.13.1/libexec/src/reflect/value.go:321 +0xb4\ngithub.com/0xProject/0x-mesh/vendor/github.com/ethereum/go-ethereum/rpc.(*callback).call(0xc00020a9c0, 0x4e10100, 0xc000236640, 0xc000437120, 0xe, 0xc0004824e0, 0x2, 0x2, 0x0, 0x0, ...)\n\t/Users/alextowle/go/src/github.com/0xProject/0x-mesh/vendor/github.com/ethereum/go-ethereum/rpc/service.go:209 +0x2b9\ngithub.com/0xProject/0x-mesh/vendor/github.com/ethereum/go-ethereum/rpc.(*handler).runMethod(0xc00011e1b0, 0x4e10100, 0xc000236640, 0xc000238000, 0xc00020a9c0, 0xc0004824e0, 0x2, 0x2, 0x2)\n\t/Users/alextowle/go/src/github.com/0xProject/0x-mesh/vendor/github.com/ethereum/go-ethereum/rpc/handler.go:369 +0x8a\ngithub.com/0xProject/0x-mesh/vendor/github.com/ethereum/go-ethereum/rpc.(*handler).handleCall(0xc00011e1b0, 0xc000482480, 0xc000238000, 0xc000478170)\n\t/Users/alextowle/go/src/github.com/0xProject/0x-mesh/vendor/github.com/ethereum/go-ethereum/rpc/handler.go:331 +0x22c\ngithub.com/0xProject/0x-mesh/vendor/github.com/ethereum/go-ethereum/rpc.(*handler).handleCallMsg(0xc00011e1b0, 0xc000482480, 0xc000238000, 0x4e03501)\n\t/Users/alextowle/go/src/github.com/0xProject/0x-mesh/vendor/github.com/ethereum/go-ethereum/rpc/handler.go:298 +0x1f3\ngithub.com/0xProject/0x-mesh/vendor/github.com/ethereum/go-ethereum/rpc.(*handler).handleMsg.func1(0xc000482480)\n\t/Users/alextowle/go/src/github.com/0xProject/0x-mesh/vendor/github.com/ethereum/go-ethereum/rpc/handler.go:139 +0x46\ngithub.com/0xProject/0x-mesh/vendor/github.com/ethereum/go-ethereum/rpc.(*handler).startCallProc.func1(0xc00011e1b0, 0xc0004a8120)\n\t/Users/alextowle/go/src/github.com/0xProject/0x-mesh/vendor/github.com/ethereum/go-ethereum/rpc/handler.go:226 +0x119\ncreated by github.com/0xProject/0x-mesh/vendor/github.com/ethereum/go-ethereum/rpc.(*handler).startCallProc\n\t/Users/alextowle/go/src/github.com/0xProject/0x-mesh/vendor/github.com/ethereum/go-ethereum/rpc/handler.go:222 +0x66\n","time":"2019-12-03T21:57:09-08:00"}
